Move ts functions to utils
