+++ cascade_ui/web/src/utils/GetVersionInfo.ts
